feat: Allow user to flag videos
User can flag a video by specifying video id. They can specify a flag
reason (else it will be "Not supplied" by default). If the video to be
flagged is the current video being played then it is stopped before it's
flagged (even if it's paused). Once a video is flagged it can no longer
be played or added to a playlist (a warning is shown if this is
attempted). Furthermore, when user tries to show all videos or show the
videos in a playist any flagged videos additionally show the reason
they've been flagged (changes where made to Video class's str dunder
method to achieve this neatly). Flagged videos also don't show up in
search results, whether searching by text or by tag. Lastly, flagged
videos will not be considered when using the random play feature.